DRM/requirements.txt
flask
streamlit
Pillow
sqlite3